bsp: Provide a fixup SSDT to define ADC chip-select and the red user LED
The upstream patch queue uses less platform code than our current patch
set. Two areas that are different:

- chip-select for the ADC on SPI0 should be described via host
  controller device property in _DSD

- the red user LED can be described as gpio-leds on the EXAR GPIO
  controller

The changes to the existing ACPI table is injected via an SSDT override.
We choose to override instead of append a new SSDT because the only the
former will work with 4.4 (upstream alone would be fine with adding).
Therefore, we hijack SSDT4 and add our additional AML fragments to it.

The table is loaded from an initrd. Yocto 2.2 lacks initrd support for
bootimg-efi, the wic plugin we use (will come with 2.3). So we need to
inject the initrd with a trick, putting it into EFI/BOOT which is copied
as-is to the target by the plugin.

The whole thing can go away again once we have a firmware update
available.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>