Fixes the acpi-upgrades recipe for custom machine configuration
When using a custom machine configuration that inherits from iot2000,
this recipe causes build issues because the MACHINE variable is
different from 'iot2000'.

This patch makes the SRC_URI machine specific, so that all machines that
are compatible to 'iot2000' are copying these tables to the work
directory. Since the acpi tables are now limited via the SRC_URI to the
compatible machine, all available tables in the work directory can be
used when compiling.

Signed-off-by: Claudius Heine <ch@denx.de>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>