Fixes conflict between the 'linux-yocto.inc' file in this and poky repo
While building it was discovered that in some cases the
'linux-yocto.inc' file in this repo overlaps with the 'linux-yocto.inc'
file in the poky repo. This caused error messages about a missing
"LICENSE" variable in bitbake while parsing the recipes.

This patch renames the "linux-yocto.inc" file to
"linux-yocto-iot2000.inc" and changes all the references accordingly.

Signed-off-by: Claudius Heine <ch@denx.de>