bsp: Fix dependency of image on acpi-upgrades
Let the image properly depend on acpi-upgrades so that required
artifacts are available in time. We need an explicit dependency of the
do_bootimg on the deployment of acpi-upgrades because the cpio file is
only available after that.

Reported by a user in the support forum and reproduced at the end via a
single-threaded build.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>