bsp: Enable support for QEMU emulation
While this will not bring us all the I/O of the IOT2000, we can still
use QEMU to test basic image features.

Set the required QB variables in the machine config, specifically
defining the RAM size (1G, that of the IOT2040) and ensuring that a
second UART is present which the image uses as console. We also need a
virtio-rng device due to lacking entropy in the virtual target.

Tune the kernel config to carry virtio block, net, and hw-random
support. As compensation, we can safely drop hw-random drivers that make
no sense on the IOT2000.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>