bsp: Add CIP-based real-time kernel
This adds support for using the RT kernel corresponding to the plain
4.4.105-cip15. The kernel is selected by setting

PREFERRED_PROVIDER_virtual/kernel = "linux-cip-rt"

We currently need two extra patches. One switches the SPI controller
driver to MSI. Otherwise, there will be an interrupt storm triggered on
the INTx line shared with the UART.

The other patch prepares the chained IRQ handler of the DesignWare APB
GPIO driver for RT-full.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>