Fix wic-image.inc
It does not make sense to first only apply ".wic" FS_TYPE
if the machine is 'iot2000' but then use the ${MACHINE} variable
for the wks file since the first append only allows one machine.

So remove the machine specific append and append for all machines.
Furthermore, we have no hard dependency on systemd-boot because we
use efibootguard.

Signed-off-by: Andreas Reichel <andreas.reichel.ext@siemens.com>