bsp: Add Quark-enabling for Node.js 8.x
This brings back the first broken and then removed x87 support for the
V8 JS engine in node, in "Quark edition". So far this is only for 8.x,
but eventually porting it further forward to 10 or 12 does not seem
unrealistic anymore - at least until V8 starts demanding more than a
single allocatable FPU register like the port always implied.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>