example: Add class for alternative npm packaging
This provides an alternative to OE's npm class which permits - with
limitations - archiving and reproducible offline installation of npm
packages with their dependencies.

The basic pattern is:
 - fetch
   - fetch npm-shrinkwrap.json
   - create stub package.json with <package-name> as dependency
   - npm ci --global-style --ignore-scripts
   - cd node_modules/<package-name>
   - add dependencies to bundledDependencies of package.json
   - npm pack
   - store tarball in DL_DIR
 - compile
   - npm install --offline --only=production [--build-from-source] <tarball>
 - install
   - copy from compilation target folder into ${D}

This works fine for a bunch of node-red packages, including the
problematic node-red-node-serialport. If will not work for packages that
do not support recompilation of packaged artifacts (with pulling their
source repos) BUT require recompilation to fulfill the target machine
requirements (in our case: Quark compiler settings).

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>