example: node-red-node-intel-gpio: Switch to IOT2000-specific version
Add detection of the IOT2000-variants to node-red-node-intel-mraa 0.0.6
which allows to expose a number of features that are specific to these
devices.

Namely, this makes the user button and the red LED of the 2040
available. Furthermore, this addresses the issue that some inputs are
not interrupt-backed and, thus, require a polling-based approach to get
updates. The new version also allows to send an initial message from
input pins to their connected nodes which helps with bringing up
boolean-like networks.

All this is now based on the code pulled from github, rather than the
distributed npm package. Because the changes are significant and in
order to avoid accidental "updates" to the official version, we declare
this package as "1.0.0-iot2000".

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>