bsp: Add gpio-sch interrupt patches
This adds the back-ported patch plus some dependencies to support
interrupts for SoC-based GPIOs. Now only DI7 and DI8 remain without
interrupt support, and that is because of the I/O expander wiring on the
board.

While textually updating the patch queue, also refresh patch 6 /wrt a
tiny comment typo that the old version contained.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>