bsp: linux-cip: Add backport of entropy collection accelerators
This adds a series of 14 patches to drivers/char/random.c that
accelerate the initial entropy collection and, thus, the boot process of
the IOT2000. This delay is primarily visible when a simple image is used
(core-image-minimal e.g.) in combination with an entropy demanding
service such as OpenSSH. It is generally not visible with the example
image and haveged installed.

The patches are only included in the regular kernel for now because the
series would need customizations for the rt kernel.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>