example: Fix node-red node management by changing gpio node installation
Using an own version for node-red-node-intel-gpio in order to avoid
accidental updates of this patched node was not a good idea. It made npm
unhappy because it couldn't find this node in its revision in the
official registry. Fix that by installing it under node-red, thus even
making it unremovable. Node-red was already out of scope for manual
updates before this.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>