example: Switch to linux-libc-headers provided by poky
Poky uses 5.4 now, and forcing packages back to our 4.4-based package
starts to cause troubles.

A nasty one was openssh failing to start its sandboxed child because it
lacks syscall defines for clock_nanosleep_time64 and therefore didn't
add this to its seccomp filter. However, glibc comes with own headers
and is using this call which cases the caller to be terminated (prior to
finding out that the syscall is actually not supported).

The price is now shipping two kernel versions in source code, but it
could we worse.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>