Add Yocto layer denpendencies in bblayers.conf
Yocto layer dependencies aren't managed in this project and the
dependency issue is transparent to users because kas insure that
necessary layers are present before launching Yocto. However, it's
important that layer dependencies are managed by native Yocto tools
because not every Yocto project is using kas. So this enables as to
include these layers in other Yocto projects without dependency issues.