Fix: Image inclusion usage correction.
Images should never be bbappended (.bbappend) because image recipies are
included (with "require") in other image recipes and "require" feature don't
guarantee .bbappend files incluson which can lead to unexpected behaviour.
So the good practice is to define new recipe for each image then include
it in other images recipe if needed.